# swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "Obligatorio Taller de Front-End",
        "version": "1.0.0",
        "title": "League Web API"
    },
    "host": "https://taller-frontend.herokuapp.com",
    "basePath": "/api",
...
    "schemes": [
        "https"
    ],
    "paths": {
...
        "/match/{matchId}": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "Get match",
                "operationId": "getMatch",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "matchId",
                    "description": "Created match id",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Match"
                        }
                    }
                }
            },
            "put": {
                "tags": [
                    "Matches"
                ],
                "summary": "Finish match",
                "description": "Consistency won't be checked.",
                "operationId": "finishMatch",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "matchId",
                    "description": "Created match id",
                    "required": true,
                    "schema": {
                        "type": "string"
                    }
                }, {
                    "in": "body",
                    "name": "body",
                    "description": "Created match object",
                    "required": true,
                    "schema": {
                        "$ref": "#/definitions/Match"
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "schema": {
                            "$ref": "#/definitions/Match"
                        }
                    }
                }
            }
        },
        "/match/getAllByChampionshipId/{championshipId}": {
            "get": {
                "tags": [
                    "Matches"
                ],
                "summary": "Get matches by championship",
                "operationId": "getMatchesByChampionshipId",
                "produces": [
                    "application/json"
                ],
                "parameters": [{
                    "in": "path",
                    "name": "championshipId",
                    "description": "Championship Id",
                    "required": true,
                    "schema": {
                        "type": "array",
                        "items": {
                            "schema": {
                                "$ref": "#/definitions/Match"
                            }
                        }
                    }
                }],
                "responses": {
                    "200": {
                        "description": "Successful operation"
                    }
                }
            }
        }
    },
    "definitions": {
...
        "Match": {
            "type": "object",
            "properties": {
                "team1": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string"
                        },
                        "players": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "team2": {
                    "type": "object",
                    "properties": {
                        "id": {
                            "type": "string"
                        },
                        "players": {
                            "type": "array",
                            "items": {
                                "type": "string"
                            }
                        }
                    }
                },
                "events": {
                    "type": "array",
                    "items": {
                        "type": "object",
                        "properties": {
                            "playerId": {
                                "type": "string"
                            },
                            "minute": {
                                "type": "number"
                            },
                            "type": {
                                "type": "string",
                                "enum": ["GOAL", "OWN_GOAL", "YELLOW_CARD", "RED_CARD"],
                                "description": "GOAL, OWN_GOAL, YELLOW_CARD, RED_CARD"
                            }
                        }
                    }
                }
            },
            "xml": {
                "name": "Match"
            }
        }
    }
}